New Record - Uploading
The new record page now successfully creates a record in the
backend. Need to put in some more testing with proper mocking and what
not but it appears to be working reasonably well. There will need to
be some changes to both the front and backends as the responsibilities
aren't really correct in terms of mimetype and pagecounts. Also next
up will need to be the records page itself so that we can actually see
what we are uploading into the backend.