Record View works
Record view now properly displays the pages of the record on the
screen. Had to do some security stuff that doesn't seem super optimal
but progress. Page now accepts the base64 data of the page in order to
actually display the pages. Also forgot to upload the actual record
view component in the last commit so that is part of that now. New
record also now goes directly to the newly uploaded record when
submitting. Both need progress bars/spinners of some sort but baby steps.